utils: has_arg returns True for keyword-only arguments
When using keyword-only arguments, has_arg would previously return
False when presented with:

    def foo(*, bar=None):
        pass

    has_arg(foo, "bar")

This made composition of functions sent to workers via client.run
using functools.partial rather fragile. For example

    def foo(a, dask_worker=None):
        return dask_worker.id

    client.run(partial(foo, 1))   # works
    client.run(partial(foo, a=1)) # fails

In the latter case, the partially applied function has a signature
where dask_worker is a keyword-only argument, in the former it is a
positional argument.

Remove this source of confusion by checking the union of positional
and keyword-only arguments.