Add Jupyter Server to Dask Scheduler
This adds an option to start a Jupyter Server in the Dask Scheduler process.

This is admittedly a bit strange, but also convenient when you have a
dask scheduler running and want a quick Jupyter notebook without setting
up separate port forwarding / auth / etc..
This piggybacks on the existing Tornado server application.

This adds a new `Scheduler(jupyter=...)` keyword
and a `dask-scheduler --jupyter` flag.

It's a bit strange, but also pretty lightweight.