fix `test_worker_doesnt_await_task_completion`
When no timeout was given to `restart`, it used 4x `Client.timeout`, which is set to `distributed.comm.timeouts.connect` :facepalm:. So what used to be a 20s timeout became a 2min timeout. And that timeout is passed down into `Worker.close`, so it gives the ThreadPoolExecutor a longer time to wait.