Dask config item for always showing GPU in scheduler dashboard menu
Currently the scheduler dashboard only shows GPU in menu if there's
a GPU on the scheduler, regardless of whether workers have GPUs.
For deployments where we know in advance that we'll have GPUs we want
to monitor on workers, it would be nice to have a way to force the
scheduler dashboard to show this menu item.