Split up `decide_worker`, remove recs
This overhauls `decide_worker` into separate methods for different
cases.

More importantly, it explicitly turns `transition_waiting_processing`
into the primary dispatch mechanism for ready tasks.

All ready tasks (deps in memory) now always get recommended to
processing, regardless of whether there are any workers in the cluster,
whether the have restrictions, whether they're root-ish, etc.

`transition_waiting_processing` then decides how to handle them
(depending on whether they're root-ish or not), and calls the
appropriate `decide_worker` method to search for a worker.

If a worker isn't available, then it recommends them off to `queued` or
`no-worker` (depending, again, on whether they're root-ish and the
WORKER_SATURATION setting).

This also updates the `no-worker` state to better match `queued`.
Before, `bulk_schedule_after_adding_worker` would send `no-worker` tasks
to `waiting`, which would then send them to `processing`. This was
weird, because in order to be in `no-worker`, they should already be ready
to run (just in need of a worker). So going straight to `processing` makes
more sense than sending a ready task back to waiting.

Finally, this adds a `SchedulerState.is_rootish` helper. Not quite the
static field on a task @fjetter wants in #6922, but a step in that
direction.