first hacky pass at structural co-assignment
the linear chain traversal of the family metric isn't good. more broadly, we need to think about what states we expect tasks in families to be in.

almost certainly, family should be a data structure that's maybe populated top-down in `update_graph`. the only reason we're able to get away with it not being a structure right now is because we assume that all tasks in a family are either not scheduled, or all in processing. there are graph structures where that's probably impossible.