test_serialize_numba: Workaround issue with np.empty_like in NP 1.23
In NumPy 1.23, the strides of empty arrays are 0 instead of the item
size, due to https://github.com/numpy/numpy/pull/21477 - however,
`np.empty_like` seems to create non-zero-strided arrays from a
zero-strided empty array, and copying to the host from a device array
with zero strides fails a compatibility check in Numba.

This commit works around the issue by calling `copy_to_host()` with no
arguments, allowing Numba to create an array on the host that is
compatible with the device array - the resulting implementation is
functionally equivalent and slightly simpler, so I believe this change
could remain permanant rather than requiring a revert later.