Support arbitrary (not just ucp-level) options in environment
We cannot control (say) transport-level (uct) options when calling
ucp.init, but if the relevant UCX environment variable lives in the
environment at the time we make the call, then it will be respected.
So take any user-provided low-level settings and temporarily insert
into the environment.

Precedence (highest to lowest):

1. Externally specified environment overrides
2. High-level distributed ucx settings
3. Low-level distributed ucx.environment settings