ucx: Remove validation of high level option names
This checking was a leftover from #3515 when users could set arbitrary
options in the top-level ucx configuration slots. Now we construct the
high level options ourselves and can therefore be sure that the
configuration names we will pass to UCX are valid.

Conversely, we cannot check the names of the arbitrary ucx.environment
options because ucp.get_config() will only return UCP-level
configuration variables (and not transport-level UCT options).