Fix UCX CUDA context test
In systems that include both MIG and non-MIG devices, it is not anymore
sufficient to rely on the CUDA device index. For example, a system with
two physical GPUs with indices 0 and 1 will see GPU ordering
differently. By default `CUDA_VISIBLE_DEVICES` will list GPUs in reverse
order of performance, thus the GPU with higher performance will have
index 0, and that will not necessarily match the Bus ID. In fact, this
is true even without MIG devices, but generally we are accustomed to see
homogenous systems, where all GPUs are identical. This will also not
necessarily match what NVML, which will (apparently) respect the Bus ID
ordering.