Special-case no-worker tasks that became root-ish
Alternative to https://github.com/dask/distributed/pull/7259. I'm quite torn about which is cleaner. I'm leaning towards this because I think it's even weirder to call `decide_worker_rootish_queuing_disabled` on a root-ish task when queuing is enabled than to call `decide_worker_non_rootish` on a root-ish task.

This also feels more consistent with the philosophy of "stick with the original decision". And if root-ish were a static property, this is what would happen.