Add methods `Client.forward_logging()` and `Client.unforward_logging()`.
These allow worker tasks to seamlessly forward LogRecords handled by
arbitrary loggers in their process back to the client process, where
they may be emitted in the user's session. In other words, this is a
simple, built-in solution for log aggregation that is independent of the
user's cluster manager (if any).

The implementation works by installing a new `ForwardLoggingPlugin` on
the worker-side, which in turn adds a custom logging Handler to the
desired logger. This Handler (`_ForwardLoggingHandler`) publishes
serialized LogRecords objects to a topic that the client subscribes to
with a handler that deserializes the LogRecord objects and sends them
off to the appropriate logger in the client process.

Care should be taken not to forward very chatty loggers at granular
levels (e.g. DEBUG or TRACE) because the large number of resulting
messages will put performance pressure on the dask scheduler. Very
granular log aggregations are still best handled by an external
aggregator, but the implementation here provides a convenient,
out-of-the-box solution for simple debugging tasks and aggregation of
sparse logging output from custom tasks.