Check ContextVar in default_client
Previously default_client would iterate through the list of clients for
the most recent.  Now it checks the ContextVar first.  This means that
default_client will be more thread-aware.

This is particularly helpful when using `wait` in a context with
multiple clients in multiple threads, a workflow that shows up in
large scale XGBoost workloads.

Also, I noticed that mypy was sad with this change, which I
wouldn't quite understand.  This didn't pass pre-commit on my machine.