Create new Zstd (De)Compressor for each compression call (#7404)
Closes #7403.

The ZstdCompressor and ZstdDecompressor objects are not thread safe to
reuse. This commit re-instantiates a new object for each call to
compress/decompress and adds a test to ensure this is the case.

The tests run the compress/decompress round trip 100 times, 10 times,
and take about 500ms for each compression algo. In my testing this was
enough to reproduce issues 90% of the time, while ensuring tests didn't
take too long.

Co-authored-by: Dylan Wragge <dylan.wragge@imc.com>
Co-authored-by: Matthew Rocklin <mrocklin@gmail.com>