p2p: Add in-memory buffering for output DiskShardsBuffer
It is likely that at least some part (possibly all) of the output of
the shuffle will fit in memory. In this circumstance, we don't need to
necessarily write output shards to disk only to read them in later.

To enable this, provide (optional) in-memory buffering on the output
DiskShardsBuffer. While the total output size is less than some limit,
don't bother hitting the disk, but rather just hold in memory
references. Once too much data has arrived, block to flush these
buffers to disk.

When reading, we now might have some shards in memory, so we
concatenate these with the on-disk buffers.