Heterogeneous Dask Cluster
The newly added `HeterogeneousCluster` supports workers with
non-identical resources communicating with the same dask scheduler. The
class inherits from the existing `SpecCluster` and adds support for
running non-identical workers.

Apart from adding the implementation of heterogeneous cluster, it makes
a few changes to the `SpecCluster` class, which might be useful for
sub-classes to customize the functionality.

Closes #7887