RFC: Add Client.decorator method
I'm not sure if this is a good idea, and if it is
then it could use a better name.  Mostly I was playing with Modal and
enjoying it but wanted full Dask semantics around futures.
Docstring follows.

Decorate a function to submit tasks to Dask

This converts a normal function to instead return Dask Futures.  That
function can then be used in parallel.

This takes the same keywords as ``client.submit``

Example
-------

```python
>>> @client.decorate()
... def f(x):
...     return x + 1

>>> futures = [f(x) for x in range(10)]
>>> results = [future.result() for future in futures]
```