Limit resources in UCX tests
Some UCX tests may fail when running from gpuCI's Docker container, even
though those are not reproducible on the host where Docker is running.
This behavior probably occurs due to the limitation of resources in the
container, thus we reduce number of workers and threads to some of the
UCX tests, which will make them unlikely to fail while still testing
basic UCX functionality.