Respect average nthreads in adaptive
Previously we would assume in adaptive that workers had a single core.

Now we do a pass through all the workers to sum up the number of threads
per machine and divide the recommendation by this amount.

This is not very expensive

```python-console
In [1]: class Worker:
   ...:     def __init__(self):
   ...:         self.cpu = 4
   ...:

In [2]: workers = [Worker() for _ in range(1000)]

In [3]: %timeit sum(w.cpu for w in workers)
33.6 µs ± 170 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```