Add option to buffer shuffle outputs in memory
This just says "I know my shuffle will fit in memory, please don't
touch the disk", gated behind setting the config option
`"distributed.shuffle.p2p.stage_in_memory"` to `True`. This needs to
be done on workers, but it should be the case that everything else is
agnostic to whether workers have set it or not (it does not need to be
a collective decision).