get_ip: handle getting 0.0.0.0 (#2554)
In Fedora's package build environment, we do some stuff to ensure
no network access is possible (this is to ensure the security,
policy compliance and reproducibility of builds, and also just
so builds don't fail or go slow or run up big networking bills).

In this environment it seems like `get_ip()` returns `"0.0.0.0"`,
which is obviously not a concrete host, but `ensure_concrete_host`
goes ahead and returns what it gets from `get_ip()` without
checking. This ultimately breaks several tests because a server's
host is considered to be `"0.0.0.0"`.

So, let's have `get_ip()` use the fallback to `getaddrinfo` if
the result of the socket path is `"0.0.0.0"`, but avoid the
warning, because pytest seems to consider warnings during test
collection as errors and we hit this code during test collection
in the affected scenario.

Signed-off-by: Adam Williamson <awilliam@redhat.com>