Remove OMP_NUM_THREADS if set to empty
Generally though, sometimes our use of OMP_NUM_THREADS makes other
systems sad.  It's unfortunately somewhat difficult to turn off.  A
common approach is that people set `OMP_NUM_THREADS=""` but this doesn't
properly unset things.

I'm curious if an approach like this would be helpful.

There might be a cleaner way to do this, and this might be a bad idea.
Please feel free to reject.  It was just easy to put this up as a PR.