Fix race condition between messages and scheduler info
SpecCluster.scheduler_info might be updated by Cluster._wait_for_workers() while messages are in flight. When scaling down
this can lead to a KeyError upon trying to get the worker's name from scheduler_info if that
worker is not present in scheduler_info anymore.

As a fix, this sends both worker key and name in the message queue so that
the name doesn't need to be looked up.