Make triggering the error more likely
Monkeypatch a slight delay and opening to run other async tasks
into worker status message processing.

In tests, this made it much more likely to trigger the error: Usually first time
with five workers on a 24 core EPYC. That means hopefully much faster and no repetitions needed.

To be confirmed on a quad core laptop!