Reuse Comm objects in Scheduler.broadcast
This updates Scheduler.broadcast to reuse (i.e. *not* close) the Comm object
that sends the message to the worker. This is motivated by the UCX comms,
which are relatively expensive to create and destroy. We noticed that a
simple ``client.run`` (which uses ``Scheduler.broadcast`` internally) took
~100s of ms to complete with UCX, most of which was spent creating and destroying
Comm objects. With this change, subsequent ``client.run`` calls take <10ms.

I've provided a new config option `distributed.scheduler.reuse-broadcast-comm`
for users who want the previous behavior. But based on https://github.com/dask/distributed/pull/3766/files,
folks agreed that this was a better default behavior. That change didn't quite
get the desired behavior because of the `close=True` passed to `send_recv`.